Use new model structure
